+++ fbpcs/CHANGELOG.md
@@ -10,5 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
   - Centralize default stage flow selection (D32774676)
+  - Finish PIDService container timeout implementation (D32600348)
 
 ### Removed